apps/dependency/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
pydantic-settings==2.7.0
prometheus-client==0.21.1
opentelemetry-api==1.29.0
opentelemetry-sdk==1.29.0
opentelemetry-exporter-otlp==1.29.0
opentelemetry-instrumentation-fastapi==0.50b0
pytest==8.3.4
httpx==0.28.1
